+++ client/cli/Cargo.toml
@@ -43,6 +43,7 @@ structopt = "0.3.8"
 sc-tracing = { version = "2.0.0-rc4", path = "../tracing" }
 chrono = "0.4.10"
 parity-util-mem = { version = "0.6.1", default-features = false, features = ["primitive-types"] }
+serde = "1.0.111"
 
 [target.'cfg(not(target_os = "unknown"))'.dependencies]
 rpassword = "4.0.1"
